Fix warnings

+++ ouzel/opengl/MeshBufferOGL.cpp
@@ -73,7 +73,7 @@ namespace ouzel
             if (vertexAttributes & VERTEX_POSITION)
             {
                 glEnableVertexAttribArray(index);
-                glVertexAttribPointer(index, 3, GL_FLOAT, GL_FALSE, vertexSize, reinterpret_cast<const GLvoid*>(offset));
+                glVertexAttribPointer(index, 3, GL_FLOAT, GL_FALSE, static_cast<GLint>(vertexSize), reinterpret_cast<const GLvoid*>(offset));
                 offset += 3 * sizeof(float);
                 index++;
             }
@@ -81,7 +81,7 @@ namespace ouzel
             if (vertexAttributes & VERTEX_COLOR)
             {
                 glEnableVertexAttribArray(index);
-                glVertexAttribPointer(index, 4, GL_UNSIGNED_BYTE, GL_TRUE, vertexSize, reinterpret_cast<const GLvoid*>(offset));
+                glVertexAttribPointer(index, 4, GL_UNSIGNED_BYTE, GL_TRUE, static_cast<GLint>(vertexSize), reinterpret_cast<const GLvoid*>(offset));
                 offset += 4 * sizeof(uint8_t);
                 index++;
             }
@@ -89,7 +89,7 @@ namespace ouzel
             if (vertexAttributes & VERTEX_NORMAL)
             {
                 glEnableVertexAttribArray(index);
-                glVertexAttribPointer(index, 3, GL_FLOAT, GL_FALSE, vertexSize, reinterpret_cast<const GLvoid*>(offset));
+                glVertexAttribPointer(index, 3, GL_FLOAT, GL_FALSE, static_cast<GLint>(vertexSize), reinterpret_cast<const GLvoid*>(offset));
                 offset += 3 * sizeof(float);
                 index++;
             }
@@ -97,7 +97,7 @@ namespace ouzel
             if (vertexAttributes & VERTEX_TEXCOORD0)
             {
                 glEnableVertexAttribArray(index);
-                glVertexAttribPointer(index, 2, GL_FLOAT, GL_FALSE, vertexSize, reinterpret_cast<const GLvoid*>(offset));
+                glVertexAttribPointer(index, 2, GL_FLOAT, GL_FALSE, static_cast<GLint>(vertexSize), reinterpret_cast<const GLvoid*>(offset));
                 offset += 2 * sizeof(float);
                 index++;
             }
@@ -105,7 +105,7 @@ namespace ouzel
             if (vertexAttributes & VERTEX_TEXCOORD1)
             {
                 glEnableVertexAttribArray(index);
-                glVertexAttribPointer(index, 2, GL_FLOAT, GL_FALSE, vertexSize, reinterpret_cast<const GLvoid*>(offset));
+                glVertexAttribPointer(index, 2, GL_FLOAT, GL_FALSE, static_cast<GLint>(vertexSize), reinterpret_cast<const GLvoid*>(offset));
                 offset += 2 * sizeof(float);
                 index++;
             }
